refactor the aog.h to separate the operation out
