+++ core/src/utility.cpp
@@ -1,0 +1,7 @@
+
+#include "utility.h"
+
+void utility::indent(int n, std::ostream &os){
+    for (int i = 0; i < n; i++)
+        os << "  ";
+}
